langchain/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.2",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "langchain-community>=0.3.24",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.21",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "wikipedia>=1.4.0",
]
